server refactored, joi added
