+++ server/const.js
@@ -1,0 +1,3 @@
+const secret = 'owieuricmudfh';
+
+module.exports = secret;
