+++ server/models/db.js
@@ -1,0 +1,16 @@
+const users = [
+  {
+    id: 0,
+    nick_name: 'Dima',
+    email: 'dima@gmail.com',
+    password: 'helloDima6',
+  },
+  {
+    id: 1,
+    nick_name: 'V',
+    email: 'V@gmail.com',
+    password: 'helloV',
+  },
+];
+
+module.exports = users;
