+++ server/package.json
@@ -14,6 +14,7 @@
     "@hapi/hapi": "^20.1.5",
     "@hapi/inert": "^6.0.3",
     "hapi-auth-jwt2": "^10.2.0",
+    "joi": "^17.4.2",
     "jsonwebtoken": "^8.5.1"
   },
   "devDependencies": {
