+++ server/routes/routes.js
@@ -1,0 +1,12 @@
+const loginHandler = require('../controllers/controller.js');
+
+const loginRoute = {
+  method: 'POST',
+  path: '/login',
+  handler: (request, h) => loginHandler(request, h),
+  options: {
+    cors: true,
+  },
+};
+
+module.exports = loginRoute;
